Worktime functions edited.
Power value of the appliances re-signed accodinng to type of the
appliance. If type is 0, then the appliance is consumer and sign of
power value is (+) positive. If type is 1, then the appliance is
producer and sign of power value is (-) negative.

In addition, <tuc> variable increased by 1 when the appliance used.

 Changes to be committed:
	modified:   worktime_nonPeriodic.m
	modified:   worktime_periodic_continuous.m
	modified:   worktime_periodic_nonContinous.m