"selectionOperator()": Added.
Selection operator used to build mate pool to generate a new population.
Chromosomes have chance to select as much as their fitness value. Of
course, there is randomization. Each of them can be selected or not
selected.

In addition there is elite team. As many as in configurations, elites
will be selected.

 Changes to be committed:
	new file:   selectionOperator.m


Former-commit-id: 31151c605038e3650cf55b23a1e6dc269e9cb70f